refactor: Make the ClientSocket class interact only with the filedescriptor
Now, the filedescriptor of the sockets are protected, so, the code which
uses it, does not have access to the file descriptors, so, its not
possible to close them directly.

Also, the database only writes and stores the content.
So, it is easier to add new functions as use main memory, send to
elasticsearch.