fix: Do a while loop to request for more data while there is data available #41
We also, removed the configuration of non blocking socket for the
client. Which was added in: https://github.com/sgtcortez/EasyKeyDB/issues/53
...
With this, if the kernel could not copy the content to the socket
buffer, and there are content to be read, we block for at most 300
milliseconds. IF no data is available, we will thrown an exception ...